requirements.txt
certifi==2020.6.20
cycler==0.10.0
future==0.18.2
imageio==2.9.0
imageio-ffmpeg==0.4.2
joblib==0.17.0
kiwisolver==1.2.0
matplotlib==3.3.2
numpy
pandas==1.1.3
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
six==1.15.0
threadpoolctl==2.1.0
torch==1.6.0
torchvision==0.7.0
